requirements.txt
asgiref==3.7.1
boto3==1.26.156
botocore==1.29.156
crispy-bootstrap5==0.7
Django==4.2.1
django-appconf==1.0.5
django-crispy-forms==2.0
django-extensions==3.2.3
django-imagekit==4.1.0
django-storages==1.13.2
django-widget-tweaks==1.4.12
Faker==18.11.2
jmespath==1.0.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
pilkit==2.0
Pillow==9.5.0
protobuf==3.20.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16

django-filter~=23.2